refactor(cargo): divide resolver into focused modules
